Restructuring integration tests to allow them to be run against existing eventstore
