+++ infra/eventstore/docker-compose.yaml
@@ -1,0 +1,18 @@
+version: '3.1'
+
+services:
+  eventstore:
+    image: eventstore/eventstore:21.10.2-buster-slim
+    restart: always
+    ports:
+      - 1113:1113
+      - 2113:2113
+    environment:
+      EVENTSTORE_CLUSTER_SIZE: 1
+      EVENTSTORE_RUN_PROJECTIONS: All
+      EVENTSTORE_START_STANDARD_PROJECTIONS: "true"
+      EVENTSTORE_EXT_TCP_PORT: 1113
+      EVENTSTORE_HTTP_PORT: 2113
+      EVENTSTORE_INSECURE: "true"
+      EVENTSTORE_ENABLE_EXTERNAL_TCP: "true"
+      EVENTSTORE_ENABLE_ATOM_PUB_OVER_HTTP: "true"
