+++ infra/run-integration-tests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+export TERM="dumb"
+./gradlew integrationTest
