+++ sourcerer-esjc-tests/build.gradle
@@ -1,10 +1,10 @@
 apply plugin: "kotlin"
 
 dependencies {
-    testImplementation project(':sourcerer-core')
-    testImplementation project(':sourcerer-esjc')
-    testImplementation project(':sourcerer-eventstore-test-utils')
-    implementation "com.github.msemys:esjc:$escjVersion"
-    testImplementation 'com.fasterxml.jackson.core:jackson-databind:2.13.2.2'
-    testImplementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.13.2'
+    integrationTestImplementation project(':sourcerer-core')
+    integrationTestImplementation project(':sourcerer-esjc')
+    integrationTestImplementation project(':sourcerer-eventstore-test-utils')
+    integrationTestImplementation "com.github.msemys:esjc:$escjVersion"
+    integrationTestImplementation 'com.fasterxml.jackson.core:jackson-databind:2.13.2.2'
+    integrationTestImplementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.13.2'
 }
