+++ sourcerer-esjc-tests/src/integrationTest/kotlin/org/elder/sourcerer/esjc/tests/EsjcEventRepositoryIntegrationTest.kt
@@ -1,0 +1,11 @@
+package org.elder.sourcerer.esjc.tests
+
+import org.elder.sourcerer.EventRepositoryFactory
+import org.elder.sourcerer.eventstore.tests.EventRepositoryIntegrationTestBase
+
+class EsjcEventRepositoryIntegrationTest
+    : EventRepositoryIntegrationTestBase() {
+    override fun createRepositoryFactory(sessionId: String): EventRepositoryFactory {
+        return createEsjcRepositoryFactory(sessionId)
+    }
+}
