+++ sourcerer-esjc-tests/src/integrationTest/kotlin/org/elder/sourcerer/esjc/tests/EsjcEventStreamsIntegrationTest.kt
@@ -1,0 +1,10 @@
+package org.elder.sourcerer.esjc.tests
+
+import org.elder.sourcerer.EventRepositoryFactory
+import org.elder.sourcerer.eventstore.tests.EventStreamsIntegrationTestBase
+
+class EsjcEventStreamsIntegrationTest : EventStreamsIntegrationTestBase() {
+    override fun createRepositoryFactory(sessionId: String): EventRepositoryFactory {
+        return createEsjcRepositoryFactory(sessionId)
+    }
+}
