+++ sourcerer-eventstore-grpc-tests/src/integrationTest/kotlin/org/elder/sourcerer/eventstoredb/GrpcEventRepositoryIntegrationTest.kt
@@ -1,0 +1,10 @@
+package org.elder.sourcerer.eventstoredb
+
+import org.elder.sourcerer.EventRepositoryFactory
+import org.elder.sourcerer.eventstore.tests.EventRepositoryIntegrationTestBase
+
+class GrpcEventRepositoryIntegrationTest : EventRepositoryIntegrationTestBase() {
+    override fun createRepositoryFactory(sessionId: String): EventRepositoryFactory {
+        return createGrpcRepositoryFactory(sessionId)
+    }
+}
