+++ sourcerer-eventstore-test-utils/build.gradle
@@ -2,6 +2,7 @@ apply plugin: "kotlin"
 
 dependencies {
     api project(':sourcerer-core')
+    implementation project(':sourcerer-kotlin')
     implementation 'junit:junit:4.12'
     implementation 'org.hamcrest:hamcrest-all:1.3'
     implementation 'org.mockito:mockito-core:1.10.19'
