+++ sourcerer-eventstore-test-utils/src/main/kotlin/org/elder/sourcerer/eventstore/tests/utils.kt
@@ -1,0 +1,7 @@
+package org.elder.sourcerer.eventstore.tests
+
+import java.util.UUID
+
+fun randomSessionId(): String {
+    return UUID.randomUUID().toString().replace("-", "")
+}
