+++ sourcerer-kotlin/build.gradle
@@ -4,9 +4,4 @@ dependencies {
     api project(':sourcerer-esjc')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.6.21"
-    testImplementation "com.github.msemys:esjc:$escjVersion"
-    testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.hamcrest:hamcrest-all:1.3'
-    testImplementation 'org.mockito:mockito-core:1.10.19'
-    testImplementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.13.2'
 }
